+++ TapSDK/TapFriendsSDK/ITapFriends.cs
@@ -18,5 +18,33 @@ namespace TapTap.Friends
         void GetFollowerList(int from, int limit, Action<List<TapUserRelationShip>, TapError> action);
         
         void GetBlockList(int from, int limit, Action<List<TapUserRelationShip>, TapError> action);
+
+        void SearchUser(string userId, Action<TapUserRelationShip, TapError> action);
+
+        void GenerateFriendInvitation(Action<string, TapError> action);
+
+        void SendFriendInvitation(Action<bool, TapError> action);
+
+        void RegisterMessageListener(ITapMessageListener listener);
+    }
+    
+    public interface ITapMessageListener
+    {
+        void OnMessageWithCode(int code, Dictionary<string, object> extras);
     }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+