+++ Assets/TapTap/Friends/CHANGELOG.md
@@ -1,5 +1,30 @@
 # ChangeLog
 
+## 2.1.2
+
+### Feature
+* 新增消息回调接口
+  ```c#
+  TapFriends.RegisterMessageListener(ITapMessageListener listener);
+  ```  
+* 搜索好友
+    ```c#
+    TapFriends.SearchUser(string userId, Action<TapUserRelationShip, TapError> action)
+    ```  
+* 分享好友邀请
+  ```c#
+  TapFriends.SendFriendInvitation(Action<bool, TapError> action);
+  ```
+* 获取好友邀请链接
+  ```c#
+  TapFriends.GenerateFriendInvitation(Action<string, TapError> action);
+  ```
+
+### Dependencies
+
+* TapTap.Bootstrap v2.1.2
+* TapTap.Common v2.1.2
+
 ## 2.1.1
 
 ### Dependencies